Append logIn route through Laravel (FileEventStore is bound as EventStore)
