+++ app/Http/Providers/AppServiceProvider.php
@@ -11,7 +11,7 @@ class AppServiceProvider extends ServiceProvider {
 	 */
 	public function boot()
 	{
-		//
+		$this->app->bind('App\Infrastructure\IEventStore', 'App\Infrastructure\FileEventStore');
 	}
 
 	/**
