+++ app/Http/Providers/EventServiceProvider.php
@@ -11,8 +11,8 @@ class EventServiceProvider extends ServiceProvider {
 	 * @var array
 	 */
 	protected $listen = [
-		'event.name' => [
-			'EventListener',
+		'App\Domain\*' => [
+			'App\Infrastructure\IEventStore@storeEvent',
 		],
 	];
